requirement.txt
faiss-cpu
torch==2.6.0
torchvision==0.21.0
uvicorn==0.31.0
fastapi==0.115.0
psycopg2
sqlmodel
sqlalchemy
databases
python-dotenv